requirements.txt
# Core dependencies
transformers>=4.35.0
torch>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
pynput>=1.7.6
pyyaml>=6.0

# API clients (optional)
openai>=1.0.0
anthropic>=0.7.0

# Video processing
ffmpeg-python>=0.2.0

# Utilities
click>=8.1.0
tqdm>=4.65.0
python-dateutil>=2.8.2
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Optional: For better performance
# accelerate>=0.24.0
# bitsandbytes>=0.41.0
